fix bug on WingNamedPipeConnection close when iocp is used
Sometimes the wing_iocp_input_stream_read_async crashed the application
because it was starting a read operation when the ThreadPoolIo object
was already deallocated. This was happening after the close of
WingNamedPipeConnection. This patch keeps the ThreadPoolIo object
alive until there is someone using it. It also invalidates the handles
in the input and output streams when the connection closes it.